+++ extensions/google-calendar/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Google Calendar Changelog
+## [1.3.0] - {PR_MERGE_DATE]
+- Feat(create-event): Enable creating events with natural language duration string input.
+- Chore(deps): Added `parse-duration`.
 
 ## [1.2.1] - {PR_MERGE_DATE}